pkgmk: fix the up-to-date option (avoid unnecesarry rebuilds)
build_needed() function returns true/yes if a source is missing,
even when the target/package exists and is up-to-date. This
behaviour triggers unnecesarry rebuilds.

Because only the remote sources can be missing and
we don't want to rebuild a port just because we've deleted
some of its remote sources, this patch changes that condition
from:
    ( the source is missing OR  is newer than the target/package )
to
    ( the source exists     AND is newer than the target/package )